print pdf and dashboard add chart

+++ controllers/SellproductController.php
@@ -73,6 +73,7 @@ class SellproductController extends Controller
         //ย้ายสินค้าในตะกร้ามาเปิดรายการสั่งซื้อ
         // เอาสินค้าในตะกร้า Yii::$app->user->identity->carts
         // return $this->asJson(Yii::$app->user->identity->carts);
+
         $bill = new Bill();
         $bill->BillDate = date('Y-m-d H:i:s');
         $bill->PeoNo = Yii::$app->user->identity->id;
@@ -81,27 +82,30 @@ class SellproductController extends Controller
             // echo $cart->pNo->Product_price * $cart->quantity;
             $sum += $cart->pNo->Product_price * $cart->quantity;
         } //คำนวณบิลรายการสินค้า 
-        $bill->BillTotal = $sum + ($sum * 0.07);
-        $bill->Billvat = $sum * 0.07;
-        $bill->BillCash = 0;   //วนลูปไป
-        $bill->save();
-
-        foreach (Yii::$app->user->identity->carts as $cart) {
-            $bill_detail = new BillDetail();
-            //สร้างออปเจ็คให้รายละเอียดบิล
-            $bill_detail->pno = $cart->pNo->PNo;
-            $bill_detail->quantity = $cart->quantity;
-            $bill_detail->amount = $cart->pNo->Product_price * $cart->quantity;
-            $bill_detail->bill_id = $bill->BillNo; //กำหนดไอดีบิลเป็นบิลที่เพิ่งเปิดรายการตรง bill->save()
-            //กำหนดค่ารายละเอียดบิล
-            $bill_detail->save();
+        $bill->BillTotal = ($sum + ($sum * 0.07));
+        $bill->Billvat = ($sum * 0.07);
+        $bill->BillCash = $bill->BillTotal;   //วนลูปไป
+        if ($bill->save()) {
+            foreach (Yii::$app->user->identity->carts as $cart) {
+                $bill_detail = new BillDetail();
+                //สร้างออปเจ็คให้รายละเอียดบิล
+                $bill_detail->pno = $cart->pNo->PNo;
+                $bill_detail->quantity = $cart->quantity;
+                $bill_detail->amount = $cart->pNo->Product_price * $cart->quantity;
+                $bill_detail->bill_id = $bill->BillNo; //กำหนดไอดีบิลเป็นบิลที่เพิ่งเปิดรายการตรง bill->save()
+                //กำหนดค่ารายละเอียดบิล
+                $bill_detail->save();
+            }
             //บันทึกรายละเอียดบิล
             $product = Product::findOne($cart->pNo->PNo); //ค้นหาโปรดัก 
             $product->Product_quantity = $product->Product_quantity - $cart->quantity; //ตัดตามจำนวนในตะกร้า เอาโปรดักจำนวนมาลบตะกร้าสินค้า
             $product->save(); //บันทึก
+
+            // return $this->asJson($bill->billDetails);
+            Cart::deleteAll("userNo = " . Yii::$app->user->identity->id); //คำสั่ง delete สำหรับเคลียร์ตะกร้าสินค้าที่ cart ที่ยูเซอปัจจุบันล้อกอินอยู่
+            return $this->redirect(Yii::$app->request->referrer ?: Yii::$app->homeUrl); //ย้อนกลับ
+        } else {
+            return $this->asJson([$bill, $bill->getErrors()]);
         }
-        // return $this->asJson($bill->billDetails);
-        Cart::deleteAll("userNo = " . Yii::$app->user->identity->id); //คำสั่ง delete สำหรับเคลียร์ตะกร้าสินค้าที่ cart ที่ยูเซอปัจจุบันล้อกอินอยู่
-        return $this->redirect(Yii::$app->request->referrer ?: Yii::$app->homeUrl); //ย้อนกลับ
     }
 }
